Removed generating licenses of pom children
Because it generated one license per each module within a library, which is not required.
We only want one license per library and we do not care aobut the lirbary's dependencies.

Fixed rewriting of project url with empty values. Usually it made sense to use the last valid non-empty url.